Secures backend environment variables
Uses Github secrets for MONGODB_URI and JWT_SECRET instead of storing them in .env file.
This change ensures that sensitive information is not exposed in the repository.

Also, updates shell scripts to ensure compatibility across different operating systems by normalizing paths, which improves the reliability of the build and deployment processes.